requirements.txt
# ── Droit du Travail RAG — dépendances Python ─────────────────────────────────
# Python 3.11 requis
# Généré pour : python ingest.py && streamlit run app.py

# ── Interface utilisateur ─────────────────────────────────────────────────────
streamlit==1.41.1

# ── LangChain core ───────────────────────────────────────────────────────────
langchain==0.3.13
langchain-community==0.3.13
langchain-anthropic==0.3.5
langchain-huggingface==0.1.2
# langchain-core unpinned: langchain-anthropic 0.3.5 requires >=0.3.33

# ── LLM — Anthropic / Claude ─────────────────────────────────────────────────
anthropic==0.42.0

# ── Embeddings — Français (CamemBERT) ────────────────────────────────────────
sentence-transformers==3.3.1
sentencepiece==0.2.0          # required by CamemBERT SentencePiece tokenizer
# PyTorch CPU uniquement (allège l'image Docker et HF Spaces)
torch==2.2.1
# transformers 4.47.x regression: CamembertTokenizerFast passes from_tiktoken=True
# for a SentencePiece model. 4.44.2 is the last stable release before this bug.
transformers==4.44.2
tokenizers==0.19.1
huggingface-hub==0.24.7

# ── Vector store ─────────────────────────────────────────────────────────────
faiss-cpu==1.9.0

# ── Utilitaires ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1
numpy==1.26.4
